+++ x/consensus/types/msgs.go
@@ -10,7 +10,7 @@ import (
 var _ sdk.Msg = &MsgUpdateParams{}
 
 func (msg MsgUpdateParams) ToProtoConsensusParams() cmtproto.ConsensusParams {
-	return cmtproto.ConsensusParams{
+	cp := cmtproto.ConsensusParams{
 		Block: &cmtproto.BlockParams{
 			MaxBytes: msg.Block.MaxBytes,
 			MaxGas:   msg.Block.MaxGas,
@@ -25,4 +25,12 @@ func (msg MsgUpdateParams) ToProtoConsensusParams() cmtproto.ConsensusParams {
 		},
 		Version: cmttypes.DefaultConsensusParams().ToProto().Version, // Version is stored in x/upgrade
 	}
+
+	if msg.Abci != nil {
+		cp.Abci = &cmtproto.ABCIParams{
+			VoteExtensionsEnableHeight: msg.Abci.VoteExtensionsEnableHeight,
+		}
+	}
+
+	return cp
 }